requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
faiss-cpu
streamlit
pandas
python-dotenv
pydantic
tiktoken